update Base to mint request_id only when needed.
Modified Base class to only mint request_id if a request_id was not
provided in constructor parameters.  This is critical for Response
class behavior, which is expected to echo back the request_id
in the response. The earlier implementation would mint a new
request_id every time the Base constructor was called.